Rationalized the sormpi and bbdydefine and bbdy so that 1. bbdy calls bbdydefine so that a separate call is not needed. 2. sormpi can be called with a -2 index and just initializes the    cartesian communicator and other mpi stuff, returning myid in    case we want to know that before starting sormpi. But we don't    actually _need_ to do that.
The interface to sormpi has not changed, so prior code should not be
broken by this rationalization. However, I did have to fiddle with
bbdydecl.f because of parameter/argument issues.

mpd.hosts, it seems mush refer to the host by what it knows to be its
hostname, otherwise problems seem to arise. Hence example mpd.hosts files.