Fixed incorrect edge field, to some extent, by setting all the cij pointers on the edge to be present and to set the iregion to zero.
Corrected some resulting problems with reverse pointers. Changed them
to refer to the position in the entire array, to avoid confusion.

That required a change in mditerate so that it works properly when called
giving the base of the cij and an offset pointer to the start of a
subarray. That was a good thing to do anyway.

Discovered and solved problems arising from non-uniform meshing.
Fixed contouring to be correct.