Fix of pernicious error in padvnc arising from getfield error when the boundary is simply the mesh edge. In that case, getfield used to do incorrect interpolation there, reaching beyond the mesh boundary to get rubbish in for the interpolation. That was prevented by changing the mesh boundary region code to -1 so that the points at the mesh edge are always considered to be outside the particle region. This ensures that they are never used as the center of field interpolation, and so we never reach beyond the mesh. In the process a more subtle logic error was found in the getfield code and corrected to make it do what it ought with adjusting the center when near a boundary.
Also implemented partexamine, which does some distribution plotting on
a cartesian grid of velocity and position. This was helpful in identifying
the above problems, because it was possible to show when and where particles
were getting kicked out of the mesh.

This was a particularly important correction because the error did not show
up on gfortran, presumably because the erroneous values accessed did not
turn out to be too stupid. Only in g77 were they bad enough to give an
obvious error. Phew!