Implemented improved free boundary condition on potential. It make phi constant along a direction the tangent of whose angle to the normal at the side boundaries is v_d. At the z-boundaries it the condition is still putting the normal derivative equal to zero, somewhat crudely.
Worked on the phireinject and rhoinf calculation because it was giving
silly answers for small boxes and low Ti. The main changes are to use
a relaxation parameter Ti/(1+Ti) for the rhoinf cfactor, and to allow some
positive phirein, up to 2Ti. At present the relaxation scheme sets the
prior phirein to zero on startup. That's going to break restarting.
Fix later.

Also change the name of the phiave file to .pha, because the longer
extension is annoying.

Fix bug in phiexamine.
Removed randc.o from objects. It is not used any more.