Implementation of periodic particles. Fixing other periodicity bugs.
Finally after many wanderings settled the issues arising from changes to
bdyshare etc. The first underlying problem was a lost comma from line 72
which broke argument passing. The second underlying problem which only
showed up with multiple core mpi running, was the necessity in sormpi
to set the full boundary for all processors, otherwise the particle
dynamics are wrong (because each processor needs the full potential including
the boundary).

On the way, partaccum was upgraded. Cartreinjection needed modification to
accommodate periodicity. Partlocate does periodic relocation. ninjcalc,
rhoinfcalc.

Then also, when the particles are periodic, the domain needs to be considered
to be just to the middle of the outermost cells, and exchange of particle
weight is required. Chargetomesh is supplemented by psumperiod. pinit needs
to account for this, as well as reinject (and its face areas).

In order to implement particle weight exchange, an new mditerator was
created that is more transparent than the old mditerate approach. It
uses a "view" of an array.