Reimplementation of mditermults and mditeradd
Retired completely the mditerate routine since it is just the same
as mditerarg, but with fewer arguments.
Also developed a version of mditerarg that uses the mditerator.
It is not (yet) the default version. For one thing it is 50% slower.
There were several problems with the implementation. The hackest one
is the iLscom calculation. Still the new version works.