Restart ability without potential information implemented.
Restart can proceed without the potential information. The potential is
simply solved for prior to the next particle advance. This permits
restart on a different grid than the prior run. So particle restart
information can be used for different grids. The restart is of course
then not perfect, because the potential solve is not exactly what would
be obtained by the next step (even if the grid is unchanged because the
initial conditions of the potential solve are different.)

An additional partlocate step is necessary because if the grid is changed
then the upper (mesh-fraction) coordinates of the particles are not
correct until that is done.