Change the setting of phip to accommodate restarts
Previously phip was set from the first object, and then if a restart
occurred it was also read from the restart files. That means the name
of output files is inappropriate if the size of the object has changed
relative to the restart files. Therefore when restarting attempt to
set the phip again from the first object, after it has been read back
in coptic.