Fix restarting not to read phi if flx is unavailable
Also fix the RefManual to reflect the changes to restart control that have
been made during the last few commits.
Fix the bug in denexamine and phiexamine that arose because the full
argument list of slicegweb was not being passed, but it must be since
there is a character argument whose length the compiler puts at the end.