Fixed restart check error caused by chi and averein problem.
Restarting was not giving identical results as diagnosed by ./check
This problem started on 09-Jan-2012 commit. It occurred only for
non-zero crelax. Prior to that commit, crelax defaulted to zero
(erroneously). Afterwards it defaulted to 0.5.
The problem was with the averaging processes that used saved values
of averein, and chi in the reinjection routines. To fix, these were
put in the form of caverein and chi into the partcom commons and then
written/read with partwriteread. This fixed things. Now cartreinject
is independent of reincom.f