Change handling of excluded regions inside point charges
If a node is outside the particle region, but inside a point-charge region
(which is the case if we have concentric point charge and finite sized
grain) then do not set its charge to compensate the electron density by
setting rho=faddu. That is problematic in such a situation because the
rhoci of the point charge gets included in faddu. Instead simply set the
(ion) charge to zero.

Updates to RefManual, diagexamine, phiexamine.