Added to makefile targets for .phi testing in geometry directory
Also added the .phi files themselves, which then serve as the reference
to which we can compare future outputs. Since they exercise a number
of different values and have switches built in, they are a much
broader test than the usual make test. But they take substantially
longer to run.