Implementation of Sampled background distribution and reinjection
In pinit and cartreinject, we now use the montecarlo sampling
technique for calculating the background distribution and the
reinjection distributions. -cp sets the power of velocity to which
the collision frequency is proportional (default 0). Iterative
technique corrects Eneutral appropriately to be consistent with
the -cp setting and vd-vneutral. Same saved distributions used
for reinjection as for particle initialization.