Fixed pernicious idiffregion bug.
tallyexit was being passed the arithmetic difference between new and old
iregions. This was an error. The treatment thereafter decided whether
to document crossings as if the bits of the passed "difference" were
the indicators of objects crossed. For that purpose, the xor of the
regions is required. Correcting this stopped the detection of some
passthrough particles, which were appearing as the primary symptom of
this problem (after implementing code to detect them).