Corrected the psumtoq to exclude electrons.
The prior solution for the inside of a point-charge region when the
node is outside the particle region was to put the ion density equal
to zero. That does not make the particle region charge free. It retains
electron charges there as well as the rhoci correction for point-charge
clouds. What we really want to do is to set the effective ion density
equal to the electron density (without the rhoic correction) so that
it nulls the plasma charge in this region, but not the rhoic charge
associated with the point's shielding cloud.
Changed in chargetomesh to do the electron cancellation.