Change potentialatpoint to fix unreasonable edge potentials
For quasineutral cases the use of fillinlin to extrapolated the potential
at the region edge led to large potentials that gave unreasonable pressure
force and other problems. Therefore the potentialatpoint routine, which
is used only by the calculateforces routine for electron pressure calculation
was changed not to do filling in, but to ignore the edge and do straight
multilinear interpolation.

Also additional traps to correct unreasonable forces on the first step
were put into fluxexamine. These prevent the plotting routines from
crashing with excessively large values. The traps were changed from a
stop to a correction.