Implement setdimens, a script to set allocated dimensions.
Also modify some places where griddecl was not needed.
It's not easy to removed griddecl completely because there
are places in padvnc and the examine code where it is needed.
Therefore setdimens can be run to set it to the defaults or to
dimensions obtained from copticgeom.dat or the script command line.