Correct allocation length ixnlength of xn to be consistent.
ixnlength has to be at least the sum of the array lengths, but formerly
it was simply set to 1000.

Instead, make meshcom.f include griddecl.f in which na_i,j,k are
defined. Then define ixnlength as the sum.

This changed the common dependencies slightly and required ndims
to be set as ndims_decl instead of ndims in griddecl.f
In coptic.f and padvnc.f ndims had then to be defined separately.