Add potential to the diag output from coptic
And modify diagexamine to read it back and optionally write a vtk
file from it. Older diag files will still sort of work, except that
the last diagnostic quantity in them will be interpreted as if it
were the potential rather than what it actually is. We ought to have
a switch to prevent that but it is not implemented (yet).