Fix griddecl inclusion and other analysis bugs.
vecglx driver was updated to re-include double buffering.

Discovered a problem with the reorganization of meshcom.f to include griddecl
It is that in analysis (or any other directory) we can't include meshcom.f
unless griddecl.f is also in this directory, because when including meshcom
griddecl.f can't otherwise be found. meshcom.f includes griddecl only
after 2 Dec 13. Added a linke from analysis/ to the file.