First implementation of multiple particle species running.
Running with particle electrons demonstrated and elementary stability
boundary observed. Only the ion species is documented for flux,
but diagnostics are available for the higher species. The density
output is the total particle charge density. (True for Boltzmann electrons,
where it is the ion charge density. Boltzmann electrons aren't particles.)
RefManual needs updating, not yet done.