Implemented trace Boltzmann facility.
This allows one to specify a small amount of linearized Boltzmann response
that is equivalent to a trace amount of Boltzmann fluid (neutralized by
an equivalent amount of background uniform charge). It has the effect of
stabilizing long wavelength background fluctuations in electron-ion
runs with domains much larger than the Debye length.