Major surgery on the object intersection code for cubes.
It had become clear that there were bugs in the cube's untested code.
In addition to fixing several on the basis of testing. Rationalization
of the code was undertaken. This required separating the intersection
finding and the face index finding (ijbin). Since the cube code was not
really working properly before, I think we can safely abandon most of
the prior code that has been deactivated. inserttest is once again a
very useful tool.